# src/main/java/br/com/elvisoliveira/literatetrain/Model/Keys.java
package br.com.elvisoliveira.literatetrain.Model;

public class Keys {
    private String privateKey;
    private String publicKey;

    public String getPublicKey() {
        return publicKey;
    }

    public void setPublicKey(String publicKey) {
        this.publicKey = publicKey;
    }

    public String getPrivateKey() {
        return privateKey;
    }

    public void setPrivateKey(String privateKey) {
        this.privateKey = privateKey;
    }
}