# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "LiterateTrain",
    "version": "v0.0.1"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:1337",
      "description": ""
    }
  ],
  "paths": {
    "/calc": {
      "get": {
        "operationId": "getSingleDigits",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SingleDigit"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "addSingleDigit",
        "requestBody": {
          "content": {
            "*/*": {
              "schema": {
                "$ref": "#/components/schemas/SingleDigit"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SingleDigit"
                }
              }
            }
          }
        }
      }
    },
    "/crypt": {
      "get": {
        "operationId": "getRandomPrivateAndPublicKeys",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Keys"
                }
              }
            }
          }
        }
      }
    },
    "/user": {
      "get": {
        "operationId": "retrieveUsers",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "addUser",
        "requestBody": {
          "content": {
            "*/*": {
              "schema": {
                "$ref": "#/components/schemas/UserEncrypted"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserEncrypted"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteUser",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserEncrypted"
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "patchUser",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserEncrypted"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "SingleDigit": {
        "required": [
          "k",
          "n"
        ],
        "type": "object",
        "properties": {
          "n": {
            "type": "string"
          },
          "k": {
            "type": "integer",
            "format": "int32"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "p": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "Keys": {
        "type": "object",
        "properties": {
          "privateKey": {
            "type": "string"
          },
          "publicKey": {
            "type": "string"
          }
        }
      },
      "UserEncrypted": {
        "required": [
          "email",
          "name",
          "publicKey"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "singleDigit": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SingleDigit"
            }
          },
          "publicKey": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        }
      },
      "User": {
        "required": [
          "email",
          "name",
          "publicKey"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "singleDigit": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SingleDigit"
            }
          },
          "publicKey": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        }
      }
    }
  }
}